feat(flutter): scaffold flutter_app + Codemagic iOS workflow
Bootstrap Flutter app `natalo_petshop_flutter` (v1.0.0) sebagai jalur
pengganti Capacitor existing dengan bundle ID identik `com.natalo.petshop`.

Foundation (lib/):
- Models: product, cart_item, member_profile (+OrderSummary, OrderItem),
  feed_post (+FeedAuthor, FeedProductLink), my_feed_post
- State stores: cart_store, member_store, recently_viewed_store,
  settings_store (theme, haptics, feed quality/autoplay/mute)
- Services: api_client (+ApiException), product, member, order, feed,
  analytics_observer, app_analytics (Firebase Analytics real),
  app_crashlytics (Firebase Crashlytics real), connectivity, deep_link,
  quick_actions, home_widget, push_notification
- Widgets: app_cart_button, app_error_widget, app_lock_gate, app_motion,
  app_product_image, app_ui (PressableTile, GlassBottomBar, StatusPill,
  SoftIconTile, InfoBanner, AnimatedEntrance, LottieAsset, EmptyState,
  SkeletonList), animated_price, favorite_button, glass_surface,
  offline_banner, product_review_section, read_only_welcome_gate,
  order_tracking_timeline, feed_comment_sheet, product_peek_dialog,
  stub_screen
- Screens: 28 (home, products, cart, checkout, login, register, member.*,
  notifications, wishlist, onboarding, dst). 4 substansial, 24 stub.
- Theme: NataloColors + NataloTheme (Material 3 light + dark)
- Utils: formatters (Rupiah, tanggal), haptics, motion_prefs, read_only_mode
- Config: api_config (ApiConfig.publicSiteUrl, ApiConfig.uri builder)

Platform scaffold parity 1:1 (iOS = Android):
- Bundle ID: com.natalo.petshop di kedua platform
- Display name: "Natalo Petshop"
- Min: iOS 13.0 / Android API 23 (Marshmallow)
- Permissions: camera, mic, photos, location, biometric, speech, push
  (iOS NSUsageDescription + UIBackgroundModes; Android uses-permission)
- Deep link: applinks Universal Links iOS (Runner.entitlements) +
  intent-filter Android — keduanya point ke natalopetshop.com.
  AASA endpoint sudah ada di app/.well-known/.../route.ts.
- Icons: 21 size iOS + 5 density Android adaptive
  (flutter_launcher_icons dari resources/icon-only.png)
- Splash: light+dark+Android 12 (flutter_native_splash dari
  resources/splash.png + splash-dark.png)

Firebase:
- Reuse project natalopetshop-818c4 (sama dengan Capacitor)
- Android: google-services.json di-copy ke flutter_app/android/app/
- iOS app baru di-register otomatis via flutterfire CLI →
  GoogleService-Info.plist di ios/Runner/ + wired ke Xcode build phase
- firebase_options.dart auto-generated, di-init di main.dart pre-runApp
- AppCrashlytics + AppAnalytics sekarang real Firebase calls (no-op kalau
  Firebase init gagal supaya app boot tetap aman)
- Gradle plugins: com.google.gms.google-services +
  com.google.firebase.crashlytics

Codemagic workflow (codemagic.yaml di root):
- workflow `flutter-ios-testflight` — build .ipa, sign via App Store
  Connect API key integration `natalo-asc`, publish ke TestFlight
- Trigger: tag `flutter-ios-v*` (manual untuk build pertama)
- Bundle yang sama dengan Capacitor — saat publish ke production, Flutter
  akan replace Capacitor build di App Store record existing

flutter analyze: 0 error.

+++ flutter_app/lib/screens/feed_screen.dart
@@ -3862,8 +3862,8 @@ class _FeedProductThumb extends StatelessWidget {
 }
 
 class _FeedProductPricing {
-  final double originalPrice;
-  final double displayPrice;
+  final int originalPrice;
+  final int displayPrice;
   final bool hasPromo;
   final int discountPercent;
